publish script

+++ publish.sh
@@ -5,7 +5,15 @@ set -e
 git checkout master
 FOLDER=$TMPDIR$RANDOM
 VERSION=$(cat package.json | sed -n "s/ *\"version\": *\"\(.*\)\",/\1/p")
-echo Publishing version $VERSION
+
+echo
+echo You are going to publish the version $VERSION
+read -p "Are you sure (y/n)? " -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    exit 1
+fi
 
 grunt
 mv dist $FOLDER
@@ -13,13 +21,6 @@ mv dist $FOLDER
 git checkout gh-pages
 rm -rf dist
 mv $FOLDER ./dist
-
-read -p "\nYou are going to publish the version $VERSION. Are you sure (y/n)? " -n 1 -r
-echo
-if [[ ! $REPLY =~ ^[Yy]$ ]]
-then
-    exit 1
-fi
 
 git add dist/
 git commit -m "version $VERSION"
